+++ packages/terra-doc-template/src/terra-dev-site/test/doc-template/DefaultDocTemplate.test.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import DocTemplate from '../../../DocTemplate';
 import readme from './common/TestReadme.md';
+import { name, version } from '../../../../package.json';
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions, import/no-duplicates */
 import TestComponentExample from './common/TestComponentExample';
 import CustomPropsTable from './common/CustomPropsTable';
@@ -27,7 +28,8 @@ const Index = () => {
   return (
     <div id="DocTemplateContainer">
       <DocTemplate
-        packageName="test-package-name"
+        packageName={name}
+        packageVersion={version}
         readme={readme}
         srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-doc-template"
         propsTables={propsTables}