+++ packages/terra-doc-template/tests/wdio/npm-badge-spec.js
@@ -1,7 +1,0 @@
-Terra.describeViewports('NpmBadge', ['medium'], () => {
-  describe('Npm Badge Variants', () => {
-    before(() => browser.url('/#/raw/tests/terra-doc-template/doc-template/npm-badge'));
-
-    Terra.it.matchesScreenshot('Badge variants display correctly');
-  });
-});