+++ packages/terra-dropdown-button/src/terra-dev-site/doc/DropdownButton/DropdownButton.1.doc.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 // Component Source
 import DropdownButtonSrc from '!raw-loader!../../../../src/DropdownButton';
@@ -23,6 +23,7 @@ import BlockDropdownButtonSrc from '!raw-loader!../../../../src/terra-dev-site/d
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[