+++ packages/terra-form-checkbox/src/terra-dev-site/doc/form-checkbox/CheckboxField.2.doc.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/CheckboxField.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -21,6 +21,7 @@ import OptionalCheckboxFieldSrc from '!raw-loader!../../../../src/terra-dev-site
 const checkboxFieldExamples = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-form-checkbox/src/CheckboxField.jsx"
     examples={[