+++ packages/terra-hyperlink/src/terra-dev-site/doc/hyperlink/Hyperlink.1.doc.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 // Component Source
 import HyperlinkSrc from '!raw-loader!../../../../src/Hyperlink';
@@ -28,6 +28,7 @@ import VideoSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/Vid
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[