+++ packages/terra-progress-bar/src/terra-dev-site/doc/progress-bar/ProgressBar.1.doc.jsx
@@ -5,7 +5,7 @@ import ReadMe from '../../../../docs/README.md';
 import ProgressBarColorDocs from '../../../../docs/progress-bar-color.md';
 import ProgressBarTwoColorsDocs from '../../../../docs/progress-bar-two-colors.md';
 import ProgressBarCustomizedColorsDocs from '../../../../docs/progress-bar-customized-colors.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -27,6 +27,7 @@ import ProgressBarCustomizedColorsSrc from '!raw-loader!../../../../src/terra-de
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[