+++ packages/terra-toggle-button/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Import for package version from package.json added to doc page
 
 3.19.0 - (August 14, 2019)
 ------------------